# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/quality_inspections:
    get:
      tags:
        - QualityInspections
      summary: List QualityInspections
      description: Get a paginated list of QualityInspections with optional filtering and sorting
      operationId: listQualityInspections
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityInspectionList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - QualityInspections
      summary: Create QualityInspection
      description: Create a new QualityInspection
      operationId: createQualityInspection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQualityInspectionRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityInspection'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/quality_inspection_parameters:
    get:
      tags:
        - QualityInspectionParameters
      summary: List QualityInspectionParameters
      description: Get a paginated list of QualityInspectionParameters with optional filtering and sorting
      operationId: listQualityInspectionParameters
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityInspectionParameterList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - QualityInspectionParameters
      summary: Create QualityInspectionParameter
      description: Create a new QualityInspectionParameter
      operationId: createQualityInspectionParameter
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateQualityInspectionParameterRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/QualityInspectionParameter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    QualityInspection:
      type: object
      required:
        - id
        - company_id
        - item_id
        - inspection_type
        - sample_size
        - inspected_at
        - status
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        template_id:
          nullable: true
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        inspection_type:
          type: string
        source_type:
          nullable: true
          type: string
          maxLength: 60
        source_id:
          nullable: true
          type: string
          format: uuid
        sample_size:
          type: integer
          format: int32
        inspected_at:
          type: string
          format: date-time
        status:
          type: string
        remarks:
          nullable: true
          type: string
          maxLength: 2000
        metadata:
          type: object
...
    CreateQualityInspectionRequest:
      type: object
      required:
        - company_id
        - item_id
        - inspection_type
        - sample_size
        - inspected_at
        - status
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        template_id:
          nullable: true
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        inspection_type:
          type: string
        source_type:
          nullable: true
          type: string
          maxLength: 60
        source_id:
          nullable: true
          type: string
          format: uuid
        sample_size:
          type: integer
          format: int32
        inspected_at:
          type: string
          format: date-time
        status:
          type: string
        remarks:
          nullable: true
          type: string
          maxLength: 2000
        metadata:
          type: object
...
    QualityInspectionList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QualityInspection'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    QualityInspectionParameter:
      type: object
      required:
        - id
        - company_id
        - template_id
        - parameter_name
        - numeric
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        template_id:
          type: string
          format: uuid
        parameter_name:
          type: string
          maxLength: 140
        numeric:
          type: boolean
        min_value:
          nullable: true
          type: string
        max_value:
          nullable: true
          type: string
        spec_text:
          nullable: true
          type: string
          maxLength: 280
        metadata:
          type: object
...
    CreateQualityInspectionParameterRequest:
      type: object
      required:
        - company_id
        - template_id
        - parameter_name
        - numeric
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        template_id:
          type: string
          format: uuid
        parameter_name:
          type: string
          maxLength: 140
        numeric:
          type: boolean
        min_value:
          nullable: true
          type: string
        max_value:
          nullable: true
          type: string
        spec_text:
          nullable: true
          type: string
          maxLength: 280
        metadata:
          type: object
...
    QualityInspectionParameterList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/QualityInspectionParameter'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []